refactor: reorganize classes into proper package structure
