+++ src/main/java/com/nirmala/logsense/LogSenseApplication.java
@@ -4,8 +4,8 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class MainApp {
+public class LogSenseApplication {
     public static void main(String[] args) {
-        SpringApplication.run(MainApp.class, args);
+        SpringApplication.run(LogSenseApplication.class, args);
     }
 }
